TTWL PC Droppers
I swear I released this but probably didn't due to the Armor being so crazy
I guess I will just update these as time goes on. If you want to help with this just DM me since it is a lot of work I will add credits for anything you help with. I don't play the game enough to know what's good.
\
Slots:
15 - ARs/Launchers
16 - Melees
17 - Pistols/Shotguns
18 - SMGs/Snipers
19 - Shields/Amus/Rings
20 - Spells
21 - Armor (UNFINISHED AF [TO MANY POSSIBILITIES])